fixes + stuff

+++ models/MaskTwoStreamConvSwitch_NET.py
@@ -6,11 +6,11 @@ import torch.nn.parallel
 import functools
 from torch.autograd import Variable
 import numpy as np
-from layer_util import *
-from MaskContextAE_NET import MaskContextAE_NET
+from .layer_util import *
+from .MaskContextAE_NET import MaskContextAE_NET
 
 
-class MaskTwoStreamConvSwitch_NET(MaskContextAE_NET):
+class   MaskTwoStreamConvSwitch_NET(MaskContextAE_NET):
     def __init__(self, opt):
         super(MaskTwoStreamConvSwitch_NET, self).__init__(opt)
         # two-stream-specific arguments
@@ -134,11 +134,11 @@ class MaskTwoStreamConvSwitch_NET(MaskContextAE_NET):
         
         layers = []
         for i in range(self.num_layers+1):
-            input_dim = output_dim
+            input_dim = int(output_dim)
             if i < self.num_layers:
                 output_dim = self.dim_list[self.num_layers-i-1]
             else:
-                output_dim = input_dim/2
+                output_dim = int(input_dim/2)
             if not(skip_layers is None):
                 if i in skip_layers: input_dim *= 2
             if self.use_simpleRes:
